chore: move some files to core
