+++ package.json
@@ -32,16 +32,14 @@
     }
   },
   "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.ts"
-    ],
-    "coveragePathIgnorePatterns": [
-      "<rootDir>/_book/",
-      "<rootDir>/node_modules/",
-      "<rootDir>/src/accessor.ts",
-      "<rootDir>/src/action.ts",
-      "<rootDir>/src/index.ts"
-    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 100,
+        "functions": 100,
+        "lines": 100,
+        "statements": 100
+      }
+    },
     "moduleFileExtensions": [
       "js",
       "ts"
