--- src/__tests__/config.test.ts
@@ -1,1 +1,0 @@
-test.skip('TODO');
