+++ src/core/__tests__/browser.test.ts
