+++ src/core/__tests__/element.test.ts
