+++ src/core/__tests__/predicate.test.ts
