+++ src/core/__tests__/verification.test.ts
