+++ src/core/accessor.ts
