+++ src/core/browser.ts
