+++ src/core/element.ts
