+++ src/core/predicate.ts
